solved multiple servers issue

+++ core/unicore_http_parse.cpp
@@ -2719,7 +2719,7 @@ int unicore_http_parse_message_body ( fsm_state_t& fsm_state , unicore_buf_t *b 
    {
       
       content_len = std::atoi ( ( char * )content_length->value );
-      if ( content_len > fsm_state.mcms )
+      if ( (size_t)content_len > fsm_state.mcms )
          return 400;
       if ( content_len <= 0 )
          return 200;
